+++ organize/filters/__init__.py
@@ -6,24 +6,19 @@ from typing_extensions import Annotated
 from .created import Created
 from .date_added import DateAdded
 from .date_lastused import DateLastUsed
-
 from .duplicate import Duplicate
 from .empty import Empty
-
-# from .exif import Exif
+from .exif import Exif
 from .extension import Extension
-
-# from .filecontent import FileContent
+from .filecontent import FileContent
 from .filter import Filter
 from .hash import Hash
 from .lastmodified import LastModified
-
-# from .macos_tags import MacOSTags
+from .macos_tags import MacOSTags
 from .mimetype import MimeType
 from .name import Name
-
-# from .python import Python
-# from .regex import Regex
+from .python import Python
+from .regex import Regex
 from .size import Size
 
 FilterType = Union[
@@ -35,11 +30,17 @@ FilterType = Union[
             Duplicate,
             DateLastUsed,
             Empty,
+            Exif,
             Extension,
+            FileContent,
             Hash,
             LastModified,
+            MacOSTags,
+            MimeType,
             Name,
             Size,
+            Python,
+            Regex,
             MimeType,
         ],
         pydantic.Field(discriminator="name"),